+++ controller/src/main/java/com/linbit/linstor/core/apicallhandler/controller/CtrlVlmDfnModifyApiCallHandler.java
@@ -20,6 +20,7 @@ import com.linbit.linstor.core.apicallhandler.controller.helpers.EncryptionHelpe
 import com.linbit.linstor.core.apicallhandler.controller.helpers.PropsChangedListenerBuilder;
 import com.linbit.linstor.core.apicallhandler.controller.internal.CtrlSatelliteUpdateCaller;
 import com.linbit.linstor.core.apicallhandler.controller.utils.SatelliteResourceStateDrbdUtils;
+import com.linbit.linstor.core.apicallhandler.controller.utils.VolumeDefinitionResizeCheckUtils;
 import com.linbit.linstor.core.apicallhandler.response.ApiDatabaseException;
 import com.linbit.linstor.core.apicallhandler.response.ApiException;
 import com.linbit.linstor.core.apicallhandler.response.ApiOperation;
@@ -34,7 +35,6 @@ import com.linbit.linstor.core.identifier.ResourceName;
 import com.linbit.linstor.core.identifier.VolumeNumber;
 import com.linbit.linstor.core.objects.Resource;
 import com.linbit.linstor.core.objects.ResourceDefinition;
-import com.linbit.linstor.core.objects.StorPool;
 import com.linbit.linstor.core.objects.Volume;
 import com.linbit.linstor.core.objects.VolumeDefinition;
 import com.linbit.linstor.core.objects.VolumeDefinition.Flags;
@@ -48,9 +48,7 @@ import com.linbit.linstor.propscon.Props;
 import com.linbit.linstor.propscon.ReadOnlyProps;
 import com.linbit.linstor.stateflags.FlagsHelper;
 import com.linbit.linstor.storage.kinds.DeviceLayerKind;
-import com.linbit.linstor.storage.kinds.DeviceProviderKind;
 import com.linbit.linstor.utils.layer.LayerRscUtils;
-import com.linbit.linstor.utils.layer.LayerVlmUtils;
 import com.linbit.locks.LockGuard;
 import com.linbit.utils.Base64;
 import com.linbit.utils.PairNonNull;
@@ -74,7 +72,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.TreeSet;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.ReadWriteLock;
@@ -285,7 +282,7 @@ public class CtrlVlmDfnModifyApiCallHandler implements CtrlSatelliteConnectionLi
         {
             if (hasDeployedVolumes(vlmDfn))
             {
-                ensureShrinkingIsSupported(vlmDfn);
+                VolumeDefinitionResizeCheckUtils.ensureShrinkingIsSupported(vlmDfn);
             }
             setFlag(vlmDfn, VolumeDefinition.Flags.GROSS_SIZE);
             updateForResize = true;
@@ -298,7 +295,7 @@ public class CtrlVlmDfnModifyApiCallHandler implements CtrlSatelliteConnectionLi
             boolean shrink = diffSize < 0;
             if (shrink)
             {
-                ensureShrinkingIsSupported(vlmDfn);
+                VolumeDefinitionResizeCheckUtils.ensureShrinkingIsSupported(vlmDfn);
                 setFlag(vlmDfn, VolumeDefinition.Flags.RESIZE_SHRINK);
                 updateForResize = true;
                 notifyStlts = true;
@@ -313,9 +310,9 @@ public class CtrlVlmDfnModifyApiCallHandler implements CtrlSatelliteConnectionLi
             }
             else
             {
-                ensureAllStorPoolsHaveEnoughFreeSpace(vlmDfn, diffSize);
+                VolumeDefinitionResizeCheckUtils.ensureAllStorPoolsHaveEnoughFreeSpace(vlmDfn, diffSize);
 
-                ensureExactSizeIsUnset(vlmDfn);
+                VolumeDefinitionResizeCheckUtils.ensureExactSizeIsUnset(vlmDfn);
 
                 updateForResize = true;
                 notifyStlts = true;
@@ -331,6 +328,8 @@ public class CtrlVlmDfnModifyApiCallHandler implements CtrlSatelliteConnectionLi
         Flux<ApiCallRc> updateResponses = Flux.empty();
         if (updateForResize)
         {
+            VolumeDefinitionResizeCheckUtils.ensureNoThickLvmSnapshots(vlmDfn);
+
             Iterator<Resource> itRsc = vlmDfn.getResourceDefinition().iterateResource();
             while (itRsc.hasNext())
             {
@@ -571,131 +570,6 @@ public class CtrlVlmDfnModifyApiCallHandler implements CtrlSatelliteConnectionLi
             {
                 throw new ApiDatabaseException(exc);
             }
-        }
-    }
-
-    /**
-     * All participating layers & providers (including providers for metadata) must support shrinking, or an Exception
-     * is thrown
-     */
-    private void ensureShrinkingIsSupported(VolumeDefinition vlmDfnRef)
-    {
-        Iterator<Resource> rscIt = vlmDfnRef.getResourceDefinition().iterateResource();
-        Set<DeviceLayerKind> layerKindSet = new HashSet<>();
-        Set<DeviceProviderKind> providerKindSet = new HashSet<>();
-        while (rscIt.hasNext())
-        {
-            Resource rsc = rscIt.next();
-            layerKindSet.addAll(LayerRscUtils.getLayerStack(rsc));
-            Set<StorPool> storPools = LayerVlmUtils.getStorPools(rsc, true);
-            for (StorPool sp : storPools)
-            {
-                providerKindSet.add(sp.getDeviceProviderKind());
-            }
-        }
-
-        for (DeviceLayerKind kind : layerKindSet)
-        {
-            if (!kind.isShrinkingSupported())
-            {
-                throw new ApiRcException(
-                    ApiCallRcImpl.entryBuilder(
-                        ApiConsts.FAIL_INVLD_VLM_SIZE,
-                        "Shrinking volumes is not supported by layer '" + kind.name() + "'. " +
-                            "Volumes can only grow in size."
-                    )
-                        .setSkipErrorReport(true)
-                        .build()
-                );
-            }
-        }
-        for (DeviceProviderKind kind : providerKindSet)
-        {
-            if (!kind.isShrinkingSupported())
-            {
-                throw new ApiRcException(
-                    ApiCallRcImpl.entryBuilder(
-                        ApiConsts.FAIL_INVLD_VLM_SIZE,
-                        "Shrinking volumes is not supported by storage provider '" + kind.name() + "'. " +
-                            "Volumes can only grow in size."
-                    )
-                        .setSkipErrorReport(true)
-                        .build()
-                );
-            }
-        }
-    }
-
-    /**
-     * All participating storage pools must have at least the additional free space left.
-     * Thin pools always fulfill this requirement.
-     *
-     * NOTE:
-     * The controller can only use the estimated additional size for checking. That means, even
-     * if the controller (barely) passes this check, the satellite might still run into an issue
-     * when executing the resize with the actual additional space, which might be more than we get
-     * here.
-     */
-    private void ensureAllStorPoolsHaveEnoughFreeSpace(VolumeDefinition vlmDfnRef, long additionalSize)
-    {
-        Iterator<Resource> rscIt = vlmDfnRef.getResourceDefinition().iterateResource();
-        Set<StorPool.Key> storPoolKeySet = new TreeSet<>();
-        while (rscIt.hasNext())
-        {
-            Resource rsc = rscIt.next();
-            Set<StorPool> storPools = LayerVlmUtils.getStorPools(rsc, true);
-            for (StorPool sp : storPools)
-            {
-                if (!sp.getDeviceProviderKind().usesThinProvisioning())
-                {
-                    if (sp.getFreeSpaceTracker().getFreeCapacityLastUpdated().orElse(0L) < additionalSize)
-                    {
-                        storPoolKeySet.add(new StorPool.Key(sp));
-                    }
-                }
-            }
-        }
-
-        if (!storPoolKeySet.isEmpty())
-        {
-            StringBuilder sb = new StringBuilder();
-            for (StorPool.Key key : storPoolKeySet)
-            {
-                sb.append("Node: ").append(key.getNodeName().displayValue)
-                    .append(", StorPool: ").append(key.getStorPoolName().displayValue).append("\n");
-            }
-            sb.setLength(sb.length() - 1); // cut last \n
-
-            throw new ApiRcException(
-                ApiCallRcImpl.simpleEntry(
-                    ApiConsts.FAIL_NOT_ENOUGH_FREE_SPACE,
-                    "Cannot grow the volume definition by " + additionalSize +
-                        "KiB, as the following storage pool do not have enough free space:\n" + sb
-                )
-            );
-        }
-    }
-
-    /**
-     * Throws an ApiRcException if the given vlmDfn's rscDfn has the "DrbdOption/ExactSize" property set to true
-     *
-     */
-    private void ensureExactSizeIsUnset(VolumeDefinition vlmDfnRef)
-    {
-        ReadOnlyProps rscDfnProps = vlmDfnRef.getResourceDefinition().getProps();
-        @Nullable String exactSize = rscDfnProps.getProp(
-            ApiConsts.KEY_DRBD_EXACT_SIZE,
-            ApiConsts.NAMESPC_DRBD_OPTIONS
-        );
-        if (exactSize != null && Boolean.parseBoolean(exactSize))
-        {
-            throw new ApiRcException(
-                ApiCallRcImpl.simpleEntry(
-                    ApiConsts.FAIL_INVLD_PROP,
-                    "Volume definition must not be resized while the resource-definition has the property '' set!",
-                    true
-                )
-            );
         }
     }