+++ server/src/main/java/com/linbit/linstor/storage/LvmDriverKind.java
@@ -49,7 +49,7 @@ public class LvmDriverKind implements StorageDriverKind
     @Override
     public boolean isSnapshotSupported()
     {
-        return false;
+        return true;
     }
 
     @Override